Reverse proxies & Inconsistency
This wiki is a cheat sheet. It is a result of analysis of various reverse proxies, cache proxies, load balancers, etc. The article (https://www.acunetix.com/blog/articles/a-fresh-look-on-reverse-proxy-related-attacks/) describes the goals of the research and how you can use the cheat sheet.
More examples can be found in the presentation (https://speakerdeck.com/greendog/reverse-proxies-and-inconsistency).